Update README.md
I went through the directions here and found a few minor issues:

Perhaps 

"git checkout -b workshop" does not work; this generates a new branch, but the `workshop` branch already exists on origin.  It should be "git checkout workshop".

It appears that the release executable is named "utxo", not "utxo-runtime", and the `--dev` flag is not recognized.

I added a few instructions on where to perform certain actions.